Refactored function directory structure
