+++ grade-management-new/GradeManagement.QueueFunction/Functions/AssignmentAcceptedEventFunction.cs
@@ -1,12 +1,10 @@
-using Azure.Storage.Queues.Models;
-
 using GradeManagement.Bll;
 using GradeManagement.Shared.Dtos.AssignmentEvents;
 
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
 
-namespace GradeManagement.QueueFunction;
+namespace GradeManagement.QueueFunction.Functions;
 
 public class AssignmentAcceptedEventFunction
 {
